+++ src/plugins/particles/util/ParticleInformationApplet.cpp
@@ -160,9 +160,9 @@ void ParticleInformationApplet::updateInformationDisplay()
 	else if(_inputMode->_pickedParticles.size() >= 2) {
 		stream << QStringLiteral("<b>") << tr("Distances:") << QStringLiteral("</b>");
 		stream << QStringLiteral("<table border=\"0\">");
-		for(int i = 0; i < _inputMode->_pickedParticles.size(); i++) {
+		for(size_t i = 0; i < _inputMode->_pickedParticles.size(); i++) {
 			const auto& p1 = _inputMode->_pickedParticles[i];
-			for(int j = i + 1; j < _inputMode->_pickedParticles.size(); j++) {
+			for(size_t j = i + 1; j < _inputMode->_pickedParticles.size(); j++) {
 				const auto& p2 = _inputMode->_pickedParticles[j];
 				stream << QStringLiteral("<tr><td>(") <<
 						(p1.particleIndex+1) << QStringLiteral(",") << (p2.particleIndex+1) <<
@@ -174,12 +174,12 @@ void ParticleInformationApplet::updateInformationDisplay()
 	if(_inputMode->_pickedParticles.size() >= 3) {
 		stream << QStringLiteral("<b>") << tr("Angles:") << QStringLiteral("</b>");
 		stream << QStringLiteral("<table border=\"0\">");
-		for(int i = 0; i < _inputMode->_pickedParticles.size(); i++) {
+		for(size_t i = 0; i < _inputMode->_pickedParticles.size(); i++) {
 			const auto& p1 = _inputMode->_pickedParticles[i];
-			for(int j = 0; j < _inputMode->_pickedParticles.size(); j++) {
+			for(size_t j = 0; j < _inputMode->_pickedParticles.size(); j++) {
 				if(j == i) continue;
 				const auto& p2 = _inputMode->_pickedParticles[j];
-				for(int k = j + 1; k < _inputMode->_pickedParticles.size(); k++) {
+				for(size_t k = j + 1; k < _inputMode->_pickedParticles.size(); k++) {
 					if(k == i) continue;
 					const auto& p3 = _inputMode->_pickedParticles[k];
 					Vector3 v1 = p2.localPos - p1.localPos;
