+++ src/plugins/pyscript/binding/PythonBinding.h
@@ -213,8 +213,8 @@ public:
 		if(i.check()) {
 			long index = i();
 			if(index < 0)
-				index += container.size();
-			if(index >= container.size() || index < 0) {
+				index += (long)container.size();
+            if(index >= (long)container.size() || index < 0) {
 				PyErr_SetString(PyExc_IndexError, "Index out of range");
 				throw_error_already_set();
 			}
